Combine AvatarCutout and PresenceCutout in to one struct (#923)
* Combine AvatarCutout and PresenceCutout in to one struct

* Move CircleCutout to its own file

* Update AvatarGroup to use .clipShape instead of .mask

* Clean up CircleCutout

* Fix indentation around presence cutout
